backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "henrylong", email = "henrylong719@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.129.1",
    "pwdlib[argon2,bcrypt]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "supabase>=2.28.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "ruff>=0.15.2",
]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = [".venv"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
